+++ CHANGELOG.md
@@ -4,7 +4,11 @@
 
 ### Other changes
 
-   - Update version to 0.1a1.dev7
+   - Update version to 0.1a1.dev7 ([f9190b87c11822203511b8784e5f4248313a59e4](https://github.com/hearot/pyrubrum/commit/f9190b87c11822203511b8784e5f4248313a59e4))
+
+### ‼️ Breaking changes
+
+   - Add `pyrubrum.menus.styles` and the `style` argument
 
 ## v0.1a1.dev6 - 2020-06-14